fix : fixing the whole pannel including header and fotter and seprating components
