+++ src/views/Task6LayoutContent.vue
@@ -1,18 +1,7 @@
 <template>
-  <div class="h-screen flex flex-col bg-gray-50">
-    <!-- Scrollable content -->
-    <main class="flex-1 overflow-y-auto p-4">
-      <div class="space-y-2">
-        <div
-          v-for="i in 400"
-          :key="i"
-          class="p-3 bg-white rounded-lg shadow hover:bg-blue-50 transition-colors"
-        >
-          Item {{ i }}
-        </div>
-      </div>
-    </main>
-  </div>
+  <main><Task6 /></main>
 </template>
 
-<script setup></script>
+<script setup>
+import Task6 from "../components/tasks/Task6.vue";
+</script>
